backend/requirements.txt
fastapi
uvicorn
snowflake-connector-python[pandas]
pandas
pyarrow
python-dotenv
